feat: enhance settings UI with comprehensive configuration options
- Move banner configuration outside of advanced settings for better visibility
- Add logging section with max_file_size_mb, retained_file_count, and directory opener
- Add extensions section for video_extensions and pdf_extensions configuration
- Fix timeout configuration to match actual config.json structure
- Add comprehensive banner configuration with template and filename patterns
- Improve CSS styling and section visibility controls
- Restore all missing functionality from original main.ts implementation

This addresses the following improvements:
- Banner section is now always visible when config is loaded
- Timeout values display correctly in advanced configuration
- Logging configuration with directory access button
- File extensions management for videos and PDFs
- Enhanced banner configuration with JSON editors for templates and patterns
