+++ .vscode/settings.json
@@ -19,6 +19,7 @@
         "danielshue",
         "fullpath",
         "LASTEXITCODE",
+        "NOTEBOOKAUTOMATION",
         "onedrive",
         "Shue",
         "subfolders"
